requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
svgwrite==1.4.1
requests==2.26.0
PyGithub
pybadges==3.0.1

